+++ src/main/java/gregtech/api/capability/impl/EnergyContainerBatteryBuffer.java
@@ -8,23 +8,15 @@ import gregtech.api.capability.IElectricItem;
 import gregtech.api.capability.IEnergyContainer;
 import gregtech.api.metatileentity.MTETrait;
 import gregtech.api.metatileentity.MetaTileEntity;
-import gregtech.common.ConfigHolder;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.items.IItemHandlerModifiable;
 
 import javax.annotation.Nullable;
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.ToLongFunction;
 
-import static gregtech.api.util.GTUtility.*;
-
-public class EnergyContainerBatteryBuffer extends MTETrait implements IEnergyContainer.IEnergyContainerOverflowSafe {
+public class EnergyContainerBatteryBuffer extends MTETrait implements IEnergyContainer {
 
     private final int tier;
 
@@ -38,13 +30,7 @@ public class EnergyContainerBatteryBuffer extends MTETrait implements IEnergyCon
         long initialAmperage = amperage;
         if(side == null || inputsEnergy(side)) {
             if (voltage > getInputVoltage()) {
-                BlockPos pos = metaTileEntity.getPos();
-                metaTileEntity.getWorld().setBlockToAir(pos);
-                if (ConfigHolder.doExplosions) {
-                    metaTileEntity.getWorld().createExplosion(null,
-                        pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5,
-                        getTierByVoltage(voltage), true);
-                }
+                IEnergyContainer.doOvervoltageExplosion(metaTileEntity, voltage);
                 return Math.min(amperage, getInputAmperage());
             }
             IItemHandlerModifiable inventory = getInventory();
@@ -107,49 +93,30 @@ public class EnergyContainerBatteryBuffer extends MTETrait implements IEnergyCon
         }
     }
 
-    private List<IElectricItem> getElectricItems() {
-        List<IElectricItem> electricItems = new ArrayList<>();
-        IItemHandlerModifiable inventory = getInventory();
-        for(int i = 0; i < inventory.getSlots(); i++) {
-            ItemStack batteryStack = inventory.getStackInSlot(i);
-            IElectricItem electricItem = getBatteryContainer(batteryStack);
-            if (electricItem != null) electricItems.add(electricItem);
-        }
-        return electricItems;
-    }
-
-    private long getCastedSum(ToLongFunction<IElectricItem> toLong) {
-        return castedSum(getElectricItems().stream()
-            .mapToLong(toLong)
-            .toArray());
-    }
-
-    private BigInteger getActualSum(ToLongFunction<IElectricItem> toLong) {
-        return sum(getElectricItems().stream()
-            .mapToLong(toLong)
-            .toArray());
-    }
-
-    private final ToLongFunction<IElectricItem> getEnergyStored = electricItem -> electricItem.discharge(Long.MAX_VALUE, getTier(), true, true, true);
-
     @Override
     public long getEnergyCapacity() {
-        return getCastedSum(IElectricItem::getMaxCharge);
-    }
-
-    @Override
-    public BigInteger getEnergyCapacityActual() {
-        return getActualSum(IElectricItem::getMaxCharge);
+        long energyCapacity = 0L;
+        IItemHandlerModifiable inventory = getInventory();
+        for (int i = 0; i < inventory.getSlots(); i++) {
+            ItemStack batteryStack = inventory.getStackInSlot(i);
+            IElectricItem electricItem = getBatteryContainer(batteryStack);
+            if (electricItem == null) continue;
+            energyCapacity += electricItem.getMaxCharge();
+        }
+        return energyCapacity;
     }
 
     @Override
     public long getEnergyStored() {
-        return getCastedSum(getEnergyStored);
-    }
-
-    @Override
-    public BigInteger getEnergyStoredActual() {
-        return getActualSum(getEnergyStored);
+        long energyStored = 0L;
+        IItemHandlerModifiable inventory = getInventory();
+        for (int i = 0; i < inventory.getSlots(); i++) {
+            ItemStack batteryStack = inventory.getStackInSlot(i);
+            IElectricItem electricItem = getBatteryContainer(batteryStack);
+            if (electricItem == null) continue;
+            energyStored += electricItem.getCharge();
+        }
+        return energyStored;
     }
 
     @Override