+++ src/main/java/gregtech/common/ClientProxy.java
@@ -5,12 +5,17 @@
 
 package gregtech.common;
 
+import codechicken.lib.reflect.ObfMapping;
+import codechicken.lib.reflect.ReflectionManager;
 import codechicken.lib.texture.TextureUtils;
+import com.mojang.authlib.minecraft.MinecraftProfileTexture.Type;
 import com.mojang.realmsclient.gui.ChatFormatting;
+import gregtech.api.GTValues;
 import gregtech.api.render.MetaTileEntityRenderer;
 import gregtech.api.unification.OreDictUnifier;
 import gregtech.api.unification.material.type.Material;
 import gregtech.api.unification.stack.UnificationEntry;
+import gregtech.api.util.GTLog;
 import gregtech.common.blocks.*;
 import gregtech.common.blocks.surfacerock.BlockSurfaceRock;
 import gregtech.common.blocks.surfacerock.BlockSurfaceRockFlooded;
@@ -20,24 +25,44 @@ import gregtech.common.render.FluidPipeRenderer;
 import gregtech.common.render.StoneRenderer;
 import net.minecraft.block.BlockColored;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.entity.AbstractClientPlayer;
+import net.minecraft.client.network.NetworkPlayerInfo;
 import net.minecraft.client.renderer.color.IBlockColor;
 import net.minecraft.client.renderer.color.IItemColor;
 import net.minecraft.item.EnumDyeColor;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.biome.BiomeColorHelper;
 import net.minecraftforge.client.event.ModelRegistryEvent;
+import net.minecraftforge.client.event.RenderPlayerEvent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.*;
+import java.nio.charset.StandardCharsets;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
+
 @SideOnly(Side.CLIENT)
 @Mod.EventBusSubscriber(Side.CLIENT)
 public class ClientProxy extends CommonProxy {
+
+    private static final ObfMapping PLAYER_INFO_MAPPING = new ObfMapping("net/minecraft/client/entity/AbstractClientPlayer", "playerInfo");
+    private static final ObfMapping PLAYER_TEXTURES_MAPPING = new ObfMapping("net/minecraft/client/network/NetworkPlayerInfo", "playerTextures");
+    private static final ResourceLocation GREGTECH_CAPE_TEXTURE = new ResourceLocation(GTValues.MODID, "textures/gregtechcape.png");
 
     public static final IBlockColor COMPRESSED_BLOCK_COLOR = (IBlockState state, IBlockAccess worldIn, BlockPos pos, int tintIndex) ->
         state.getValue(((BlockCompressed) state.getBlock()).variantProperty).materialRGB;
@@ -105,6 +130,7 @@ public class ClientProxy extends CommonProxy {
     public void onPostLoad() {
         super.onPostLoad();
         MetaTileEntityRenderer.postInit();
+        startCapeLoadingThread();
     }
 
     public void registerColors() {
@@ -130,6 +156,69 @@ public class ClientProxy extends CommonProxy {
                     event.getToolTip().add(1, ChatFormatting.GRAY.toString() + unificationEntry.material.chemicalFormula);
                 }
             }
+        }
+    }
+
+    private static final Set<UUID> capeHoldersUUIDs = new HashSet<>();
+
+    private static void startCapeLoadingThread() {
+        Thread capeListLoadThread = new Thread(ClientProxy::loadCapesList, "GregTech Cape List Downloader");
+        capeListLoadThread.setDaemon(true);
+        capeListLoadThread.start();
+    }
+
+    private static void loadCapesList() {
+        capeHoldersUUIDs.add(UUID.fromString("4bdba267-1479-449a-8ae4-d1957dd39f29"));
+        capeHoldersUUIDs.add(UUID.fromString("6cb05251-cd1b-481e-bf59-07637add1c64"));
+        try {
+            URL connectURL = new URL("https://www.dropbox.com/s/zc07k4y1h4ftmz3/GregTechPatreonList.txt?dl=1");
+            HttpURLConnection connection = (HttpURLConnection) connectURL.openConnection(Minecraft.getMinecraft().getProxy());
+            try {
+                connection.setDoInput(true);
+                connection.setDoOutput(false);
+                connection.connect();
+                InputStream inputStream = connection.getInputStream();
+                capeHoldersUUIDs.addAll(retrieveCapeUUIDs(inputStream));
+            } finally {
+                connection.disconnect();
+            }
+        } catch (UnknownHostException |
+            SocketTimeoutException |
+            MalformedURLException ignored) {
+        } catch (IOException exception) {
+            GTLog.logger.warn("Failed to fetch cape list", exception);
+        }
+    }
+
+    private static Set<UUID> retrieveCapeUUIDs(InputStream inputStream) throws IOException {
+        InputStreamReader streamReader = new InputStreamReader(inputStream, StandardCharsets.UTF_8);
+        BufferedReader reader = new BufferedReader(streamReader);
+        Set<UUID> result = new HashSet<>();
+        for (; ; ) {
+            String line = reader.readLine();
+            if (line == null)
+                break;
+            int firstCommentIndex = line.indexOf('#');
+            if (firstCommentIndex > -1) {
+                line = line.substring(0, firstCommentIndex);
+            }
+            try {
+                UUID playerUUID = UUID.fromString(line.trim());
+                result.add(playerUUID);
+            } catch (IllegalArgumentException exception) {
+                GTLog.logger.warn("Failed to parse cape player UUID {}", line.trim(), exception);
+            }
+        }
+        return result;
+    }
+
+    @SubscribeEvent
+    public static void onPlayerRender(RenderPlayerEvent.Pre event) {
+        AbstractClientPlayer clientPlayer = (AbstractClientPlayer) event.getEntityPlayer();
+        if(capeHoldersUUIDs.contains(clientPlayer.getUniqueID()) && clientPlayer.hasPlayerInfo() && clientPlayer.getLocationCape() == null) {
+            NetworkPlayerInfo playerInfo = ReflectionManager.getField(PLAYER_INFO_MAPPING, clientPlayer, NetworkPlayerInfo.class);
+            Map<Type, ResourceLocation> playerTextures = ReflectionManager.getField(PLAYER_TEXTURES_MAPPING, playerInfo, null);
+            playerTextures.put(Type.CAPE, GREGTECH_CAPE_TEXTURE);
         }
     }